+++ FE/src/components/main/Main.jsx
@@ -6,8 +6,8 @@ import IssueList from './IssueList';
 import DropDownFilter from './DropDownFilter';
 import { useFilterContext } from '../../context/FilterContext';
 import mockIssueList from '../../data/issueList.json';
-import { StyledHeaderContents } from '../../styles/theme';
 import NavStateType from './NavStateType';
+import NavFilterType from './NavFilterType';
 
 // TODO: fetch 데이터
 const labelTypeItems = [
@@ -84,12 +84,6 @@ export default function Main() {
         return filters;
     };
 
-    const dispatchIssue = ({ target }) => {
-        const attrValue = target.getAttribute('attr-key');
-        if (!attrValue) return;
-        dispatch({ type: dispatchTypeByFilterContents[attrValue], payload: attrValue });
-    };
-
     const toggleEntireCheckBox = () => {
         if (checkedItems.length === mockIssueList.length) setCheckedItems([]);
         else setCheckedItems(mockIssueList.map(({ id }) => id));
@@ -144,46 +138,14 @@ export default function Main() {
                             dispatch={dispatch}
                         ></NavStateType>
                     ) : (
-                        <StyledHeaderContents>
-                            <div className="issue">
-                                <span
-                                    className={`issueOption click ${selectedFilters.issues.isClosed ? '' : `checked`}`}
-                                    attr-key="is:open"
-                                    onClick={dispatchIssue}
-                                >
-                                    열린 이슈()
-                                </span>
-                                <span
-                                    className={`issueOption click ${selectedFilters.issues.isClosed ? `checked` : ''}`}
-                                    attr-key="is:closed"
-                                    onClick={dispatchIssue}
-                                >
-                                    닫힌 이슈()
-                                </span>
-                            </div>
-                            <div className="filter">
-                                <span className="filterOption">
-                                    <DropDownFilter filterTitle={'author'} filterItems={imageTypeItems} dispatch={dispatch}>
-                                        담당자
-                                    </DropDownFilter>
-                                </span>
-                                <span className="filterOption">
-                                    <DropDownFilter filterTitle={'label'} filterItems={labelTypeItems} dispatch={dispatch}>
-                                        레이블
-                                    </DropDownFilter>
-                                </span>
-                                <span className="filterOption">
-                                    <DropDownFilter filterTitle={'milestone'} filterItems={milestoneTypeItems} dispatch={dispatch}>
-                                        마일스톤
-                                    </DropDownFilter>
-                                </span>
-                                <span className="filterOption">
-                                    <DropDownFilter filterTitle={'assignee'} filterItems={imageTypeItems} dispatch={dispatch}>
-                                        작성자
-                                    </DropDownFilter>
-                                </span>
-                            </div>
-                        </StyledHeaderContents>
+                        <NavFilterType
+                            dispatchTypeByFilterContents={dispatchTypeByFilterContents}
+                            imageTypeItems={imageTypeItems}
+                            labelTypeItems={labelTypeItems}
+                            milestoneTypeItems={milestoneTypeItems}
+                            dispatch={dispatch}
+                            ischecked={selectedFilters.issues.isClosed}
+                        ></NavFilterType>
                     )}
                 </StyledBoxHeader>
 
@@ -207,14 +169,6 @@ const StyledBoxBody = styled.div`
     min-height: 80px;
     /* background-color: skyblue; */
 `;
-
-// const StyledHeaderContents = styled.div`
-//     width: 95%;
-//     display: flex;
-//     flex-direction: row;
-//     justify-content: space-between;
-//     align-items: center;
-// `;
 
 const StyledBoxHeader = styled.div`
     background-color: ${(props) => props.theme.listHeaderColor};
